feat: Make publication date show in client's timezone
